Download Image Update PHPImage.php
Explanation of Changes:
Content-Type Header: The Content-Type header remains the same, specifying the MIME type of the image.
Content-Disposition Header: This is added to indicate the file should be downloaded. The filename parameter defines the name of the downloaded file.
Default Filename: If no filename is passed to the download() method, it defaults to image.